#29 refactor : care category
