+++ care-doc/src/components/Care.js
@@ -3,8 +3,7 @@ import styles from '../css/Care.module.css';
 import { useSelector, useDispatch } from 'react-redux';
 import { clearStep } from '../store/pageStep';
 import { clearType } from '../store/careType';
-import Footer1 from './Footer1';
-import Title from './Title';
+import Footer from './Footer';
 import Category from './Category';
 import Header from './Header';
 import Schedule from './Schedule';
@@ -36,9 +35,8 @@ const Care = () => {
         돌봄 유형 <span>{currentStep}</span>
         <span> / 4</span>
       </div>
-      <Title text="돌봄 유형을 설정해주세요" />
       {renderPageStep(currentStep)}
-      <Footer1 />
+      <Footer />
     </main>
   );
 };
