+++ care-doc/src/index.js
@@ -7,11 +7,9 @@ import { BrowserRouter } from 'react-router-dom';
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
 import { Provider } from 'react-redux';
 import pageStepReducer from './store/pageStep';
-import staticDataReducer from './store/staticData';
 import careTypeReducer from './store/careType';
 
 const reducers = combineReducers({
-  static: staticDataReducer,
   pageStep: pageStepReducer,
   careType: careTypeReducer,
 });
